refactor: rename package message to dialog
