+++ src/main/java/com/quickmemo/plugin/ui/dialog/CreatedMemoDialog.java
@@ -1,4 +1,4 @@
-package com.quickmemo.plugin.ui.message;
+package com.quickmemo.plugin.ui.dialog;
 
 import com.intellij.openapi.ui.Messages;
